requirements.txt
absl-py==1.4.0
altair==4.2.2
appdirs==1.4.4
asgiref==3.5.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.2.2
contourpy==1.0.7
crispy-bootstrap4==2022.1
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
distlib==0.3.5
Django==4.0.6
django-address==0.2.8
django-cors-headers==4.1.0
django-crispy-forms==1.14.0
django-phone-field==1.8.1
djangorestframework==3.14.0
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.7.1
flatbuffers==23.3.3
fonttools==4.39.3
frozendict==2.3.8
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
html5lib==1.1
idna==3.4
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.18.1
ipywidgets==8.1.2
jax==0.4.8
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyterlab_widgets==3.0.10
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.0.4
mpmath==1.3.0
multitasking==0.0.11
mysql-connector-python==8.4.0
mysqlclient==2.1.1
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.0
pandas==2.2.2
pandas-datareader==0.10.0
parso==0.8.4
pexpect==4.9.0
Pillow==9.5.0
pipenv==2022.7.4
platformdirs==2.5.2
prompt-toolkit==3.0.43
protobuf==3.20.3
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
pygame==2.1.2
Pygments==2.15.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.3
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.2
stack-data==0.6.3
streamlit==1.33.0
sympy==1.11.1
tenacity==8.2.3
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2024.2.12
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torch-summary==1.4.5
torch-tb-profiler==0.4.3
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.2
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.14
validators==0.20.0
venv-run==0.1.2
virtualenv==20.15.1
virtualenv-clone==0.5.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.14.1
yfinance==0.2.20
zipp==3.15.0
